# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Video Tools API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/video-processing-jobs": {
      "get": {
        "summary": "List processing jobs",
        "tags": [
          "processor"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "videoProjectId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "queued",
                "running",
                "done",
                "failed",
                "canceled"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "videoProjectId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "jobType": {
                            "type": "string",
                            "enum": [
                              "transcribe_asset",
                              "extract_short_clip",
                              "render_edited_master",
                              "render_short"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "queued",
                              "running",
                              "done",
                              "failed",
                              "canceled"
                            ]
                          },
                          "progress": {
                            "type": "integer"
                          },
                          "stage": {
                            "type": "string",
                            "nullable": true
                          },
                          "result": {
                            "type": "object",
                            "additionalProperties": {},
                            "nullable": true
                          },
                          "error": {
                            "type": "string",
                            "nullable": true
                          },
                          "attempts": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "startedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "finishedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "videoProjectId",
                          "jobType",
                          "status",
                          "progress",
                          "stage",
                          "result",
                          "error",
                          "attempts",
                          "createdAt",
                          "startedAt",
                          "finishedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/video-processing-jobs/{jobId}": {
      "get": {
        "summary": "Get a job's status",
        "tags": [
          "processor"
        ],
        "description": "The SPA polls this. Swappable for SSE later without changing the contract.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "jobId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "videoProjectId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "jobType": {
                      "type": "string",
                      "enum": [
                        "transcribe_asset",
                        "extract_short_clip",
                        "render_edited_master",
                        "render_short"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "queued",
                        "running",
                        "done",
                        "failed",
                        "canceled"
                      ]
                    },
                    "progress": {
                      "type": "integer"
                    },
                    "stage": {
                      "type": "string",
                      "nullable": true
                    },
                    "result": {
                      "type": "object",
                      "additionalProperties": {},
                      "nullable": true
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "attempts": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "startedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "finishedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "videoProjectId",
                    "jobType",
                    "status",
                    "progress",
                    "stage",
                    "result",
                    "error",
                    "attempts",
                    "createdAt",
                    "startedAt",
                    "finishedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/video-processing-jobs/{jobId}/cancel": {
      "post": {
        "summary": "Cancel a queued or running job",
        "tags": [
          "processor"
        ],
        "description": "Best-effort: a running job's ffmpeg child is killed — a wasted render burning CPU is worse.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "jobId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "videoProjectId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "jobType": {
                      "type": "string",
                      "enum": [
                        "transcribe_asset",
                        "extract_short_clip",
                        "render_edited_master",
                        "render_short"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "queued",
                        "running",
                        "done",
                        "failed",
                        "canceled"
                      ]
                    },
                    "progress": {
                      "type": "integer"
                    },
                    "stage": {
                      "type": "string",
                      "nullable": true
                    },
                    "result": {
                      "type": "object",
                      "additionalProperties": {},
                      "nullable": true
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "attempts": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "startedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "finishedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "videoProjectId",
                    "jobType",
                    "status",
                    "progress",
                    "stage",
                    "result",
                    "error",
                    "attempts",
                    "createdAt",
                    "startedAt",
                    "finishedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000"
    }
  ],
...
}